+++ opengever/meeting/browser/proposaloverview.py
@@ -3,10 +3,10 @@ from opengever.base.browser.helper import get_css_class
 from opengever.bumblebee import is_bumblebee_feature_enabled
 from opengever.document import _ as document_mf
 from opengever.document.widgets.document_link import DocumentLinkWidget
-from opengever.meeting.browser.proposaltransitions import ProposalTransitionController
 from opengever.meeting.interfaces import IHistory
 from opengever.meeting.model import SubmittedDocument
 from opengever.meeting.proposal import ISubmittedProposal
+from opengever.meeting.proposal_transition_comment import ProposalTransitionCommentAddForm
 from opengever.tabbedview import GeverTabMixin
 from plone.app.contentlisting.interfaces import IContentListing
 from plone.dexterity.browser import view
@@ -18,7 +18,7 @@ class OverviewBase(object):
         return self.context.get_transitions()
 
     def transition_url(self, transition):
-        return ProposalTransitionController.url_for(self.context, transition.name)
+        return ProposalTransitionCommentAddForm.url_for(self.context, transition.name)
 
     def is_comment_allowed(self):
         return self.context.can_comment()
